[typescript] Move bulk of component helpers. 300+ to go.
