+++ to.etc.domui/src/main/resources/resources/ts/domui.jquery.d.ts
@@ -10,6 +10,8 @@ interface JQuery {
 	fixOverflow() : void;
 	setBackgroundImageMarker() : void;
 
+	ColorPicker: any;
+
 }
 
 interface JQueryStatic {
